perf: use compact_str crate to reduce heap allocations for GridCell
This commit uses `CompactString` struct from [ParkMyCar/compact_str][1]
as the data type for `GridCell::contents` field. `CompactString` by
default will store smaller strings inline on the stack. On 64 bit
architectures this allows storing 24 bytes on the stack and heap
allocates if more than 24 bytes is needed.

This improves performance as most of `GridCell::contents` are 24 bytes
or smaller thus reducing the number of heap allocations.

Changes in dependencies:
* bump nls_term_grid version to "0.3.0"
* add compact_str

[1]: https://crates.io/crates/compact_str