perf: replace usages of String with CompactString
This commit uses `CompactString` struct from [ParkMyCar/compact_str][1]
as a replacement for `String` struct. `CompactString` by default will
store smaller strings inline on the stack. On 64 bit architectures this
allows storing 24 bytes on the stack and heap allocates if more than 24
bytes is needed.

This improves performance as most `String` usages are 24 bytes or
smaller thus reducing the number of heap allocations.

[1]: https://crates.io/crates/compact_str